requirements.txt
customtkinter
patool

